refactor(solver): name content predicate guard states (#14351)
